1/3/2024 diep
Update Login and Register

+++ AzCare-APIServices/src/main/java/com/project/AzCar/Repositories/Users/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.project.AzCar.Entities.Users.Users;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users,Integer>{
+public interface UserRepository extends JpaRepository<Users,Long>{
 
 	
 	public Users findByEmail(String email);
